Main Conformance: Update EXP, POW precision
EXP and POW are complex to implement in a precise number of ulp
and a more flexible precision is given.

Additionally, the test generater S=5 has the bias values
changed to zero. If the bias values are non-zero then a
bias for all output tensor elements of the channel can affect
the dot product bias test.

The numeric accuracy helpers are updated for bf16.

Signed-off-by: Dominic Symes <dominic.symes@arm.com>
Change-Id: Ia46a2ef4d577244b6983a08ce850de3db9573a42