Main Conformance: TANH change, SCATTER clarification
TANH can have cancellation between exp(x) and exp(-x)
for small abs(x). To allow for this, a lower bound of
0.5*exp(-normal_fraction<in_out_t>) is set for the
error bound. The TANH formula also can contain 4
exp() values, so the error bound is increased to 4
multiplied by the exp() error bound.

SCATTER output size K must be as least as great
as the number of indices W, as each index is distinct.

Change-Id: I8d1f1c866bd59d8140cf7db76404f5006a04a3f3
Signed-off-by: Dominic Symes <dominic.symes@arm.com>