Return the first match of NaN from ARGMAX
In propagating NaN mode, Tosa argmax as currently defined returns
the last NaN if multiple NaNs. This appears to differ from the
"first match is returned" rule that is adopted in NumPy and PyTorch.
Also the operator should consistently return the first match along
the search axis as it does for normal FP values.
Change the pseudo code of argmax to match the behavior.

Signed-off-by: TatWai Chong <tatwai.chong@arm.com>
Change-Id: I3e289071d0c53007476b7f025a9983b4c6e749b8