Improve consistency of rank 0 tensor handling
Define behavior of for_each and for_each_data_position in pseudocode.

Addresses issues where the specification was unclear on the behavior
of rank 0 tensors.

Remove support for CONCAT of rank 0 tensors.
Allow RESHAPE of rank 0 tensors.
Define behavior of for_each on a rank 0 tensor.
Allow COND_IF condition to be rank 0.

Signed-off-by: Eric Kunze <eric.kunze@arm.com>
Change-Id: I8a01954b2eeab4160901d8b71d2d6da29ba86f8a