Remove restriction on RFFT output
Rather than forcing the output to zero at (0,0), (0,W/2), (H/2,0) and
(H/2,W/2), we allow them to use the dot product accuracy check.

Implementations may choose to skip the multiply at these points, or
do a muliply of 0. If the input value is inf or a NaN, the reference value
is NaN, which makes results valid.

Signed-off-by: Eric Kunze <eric.kunze@arm.com>
Change-Id: I50a34cde414d62e36c8aceb695a1e6a5d0ee83c4