Modify RESCALE rounding options
The current RESCALE has two options, single and double rounding.
This changes to an enumeration listing options for the RESCALE
rounding mode.

SINGLE_ROUND retains the existing behavior of single rounding.

INEXACT_ROUND is a new mode that allows non bit exact behavior
when doing rounding and provides guidelines for the precision
required in that case. INEXACT_ROUND mode is only available if
the EXT-INEXACTROUND extension is supported.

The original double_round argument is now covered under the
new rounding mode and is represented as the DOUBLE_ROUND option
if the EXT-DOUBLEROUND extension is supported.

Implementations supporting DOUBLE_ROUND need to pass the existing
conformance tests for double rounding. If a TOSA graph with
rounding set to DOUBLE_ROUND is passed to an implementation that
does not support EXT-DOUBLEROUND, the graph is invalid.

This change adds a mechanism in the spec tooling to allow extensions to
add new values to enums. Extensions can add entirely new enums, or
values to an existing enum.

Signed-off-by: Eric Kunze <eric.kunze@arm.com>
Change-Id: I11035dd08d01555c4d36d16f1b1c2fab98066b15