Fix a typo in CONV2D error_if int8_t -> i8_t
Weight types are only allowed to be signless, but the current
error if check required a signed type. This commit updates the
required type to be signless, aligning with other operators e.g.
avg_pool2d, conv3d, etc.

In addition, this commit also improves the consistency of the
comment that follows these zero point error if checks across all
operators that use them.

Change-Id: I03c45e9ba1324450187aadb0e08f411e84831f63
Signed-off-by: Luke Hutton <luke.hutton@arm.com>