Clarify compatibility between profiles and extensions
All extensions require at least one specific profile to be supported.
Wording this in terms of allowed profiles leads to ambiguous cases when
implementations support multiple profiles. For example, is supporting
EXT-BF16 allowed for an implementation that supports both PRO-INT and
PRO-FP? PRO-INT was not listed as an allowed profile for EXT-BF16.

Signed-off-by: Kevin Petit <kevin.petit@arm.com>
Change-Id: I55de341189746008ca42ebbeb047ffb270cb32b0