Add a REQUIRE to DIM to check for invalid dim size (#19)
This commit adds an evaluation time check to make sure
DIM doesn't return a non-positive value. While shape_t
allows zero to be used for parameters like padding and
negative values to be "used for negative offsets" in
operations such as resize, it does not make sense for
the dim of a tensor to be zero or negative at the time
of runtime execution. Adding this check allows software
implementations to reserve negative dims in a tensor
shape for other purposes, such as representing a
dynamic dimension that is yet to be resolved.

Change-Id: I11c561d303fdb9df3a42fd3d0fa2ddd9cf663995

Signed-off-by: Luke Hutton <luke.hutton@arm.com>